Updated AGC004_A (AC)

+++ agc004_a/Main.java
@@ -4,11 +4,11 @@ import java.util.Scanner;
 class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int A = sc.nextInt();
-		int B = sc.nextInt();
-		int C = sc.nextInt();
+		long A = sc.nextLong();
+		long B = sc.nextLong();
+		long C = sc.nextLong();
 
-		int[] array = new int[] { A, B, C };
+		long[] array = new long[] { A, B, C };
 		Arrays.sort(array);
 
 		if (A % 2 == 0 || B % 2 == 0 || C % 2 == 0) {
